refactor(operator): dedupe Secret test doubles, drop test-fn doc comments
Addresses grid#59 review feedback (praxis-bot):

- mock_kube_client_with_secrets was duplicated verbatim (~40 lines)
  between secret.rs::tests and endpoint_tls.rs::tests, with
  secret_with_key also duplicated in spirit. Extracted both into a new
  resources::test_doubles module (cfg(test)-gated at the mod
  declaration) so the two copies cannot drift independently.
- Removed doc comments on
  read_secret_bytes_key_absent_from_existing_secret_returns_key_missing
  (secret.rs) and
  resolve_tls_config_ca_key_absent_from_existing_secret_yields_key_missing
  (endpoint_tls.rs) per the project convention (CLAUDE.md, Test
  Organization: function name is the documentation) -- both were
  redundant with the grid#58 regression context already carried by
  each test's assertion message.
