+++ operator/src/resources/endpoint_tls.rs
@@ -315,67 +315,8 @@ async fn read_tls_secret_for_verify(
 mod tests {
     use std::collections::HashMap;
 
-    use k8s_openapi::{ByteString, api::core::v1::Secret};
-
     use super::*;
-
-    // -----------------------------------------------------------------------
-    // Test doubles
-    // -----------------------------------------------------------------------
-
-    /// Build a `kube::Client` backed by an in-memory map of Secret name to
-    /// `Secret`, so `resolve_tls_config`/`verify_tls_accessible` can be
-    /// exercised without a real cluster. Any name not present in the map
-    /// returns HTTP 404.
-    #[expect(
-        clippy::too_many_lines,
-        reason = "test mock builder: 404-vs-200 branches are the whole point"
-    )]
-    fn mock_kube_client_with_secrets(secrets: HashMap<&'static str, Secret>) -> kube::Client {
-        let service = tower::service_fn(move |req: http::Request<kube::client::Body>| {
-            let secrets = secrets.clone();
-            async move {
-                let name = req.uri().path().rsplit('/').next().unwrap_or_default().to_owned();
-                let response = secrets.get(name.as_str()).map_or_else(
-                    || {
-                        let not_found = serde_json::json!({
-                            "kind": "Status",
-                            "apiVersion": "v1",
-                            "status": "Failure",
-                            "message": format!("secrets \"{name}\" not found"),
-                            "reason": "NotFound",
-                            "code": 404,
-                        });
-                        http::Response::builder()
-                            .status(404)
-                            .body(kube::client::Body::from(
-                                serde_json::to_vec(&not_found).unwrap_or_else(|_| std::process::abort()),
-                            ))
-                            .unwrap_or_else(|_| std::process::abort())
-                    },
-                    |secret| {
-                        http::Response::builder()
-                            .status(200)
-                            .body(kube::client::Body::from(
-                                serde_json::to_vec(secret).unwrap_or_else(|_| std::process::abort()),
-                            ))
-                            .unwrap_or_else(|_| std::process::abort())
-                    },
-                );
-                Ok::<_, std::convert::Infallible>(response)
-            }
-        });
-        kube::Client::new(service, "default")
-    }
-
-    fn secret_with_key(key: &str, value: &[u8]) -> Secret {
-        let mut data = std::collections::BTreeMap::new();
-        data.insert(key.to_owned(), ByteString(value.to_vec()));
-        Secret {
-            data: Some(data),
-            ..Default::default()
-        }
-    }
+    use crate::resources::test_doubles::{mock_kube_client_with_secrets, secret_with_key};
 
     fn test_tls_config(ca_secret_name: &str) -> EndpointTlsConfig {
         EndpointTlsConfig {
@@ -533,10 +474,6 @@ mod tests {
         assert_eq!(reason, TlsFailureReason::SecretMissing);
     }
 
-    /// Regression test for grid#58 at the `InferenceProvider`
-    /// metrics/health-check TLS resolution path (the pipeline the bug was
-    /// originally reported against): a key absent from an existing CA
-    /// Secret must surface as `KeyMissing`, not `SecretMissing`.
     #[tokio::test]
     async fn resolve_tls_config_ca_key_absent_from_existing_secret_yields_key_missing() {
         let client =
