+++ operator/src/resources/mod.rs
@@ -35,6 +35,10 @@ pub(crate) mod provider_metrics;
 pub mod routing_overlay;
 /// Secret builders for grid TLS certificates.
 pub mod secret;
+/// Shared Kubernetes Secret test doubles, reused by `secret`/`endpoint_tls`
+/// unit tests instead of each keeping its own copy of the same mock.
+#[cfg(test)]
+pub(crate) mod test_doubles;
 /// Trust bundle management for grid mTLS.
 pub mod trust_bundle;
